Refactor, moving CLI commands to their own module
Other than use of pytest's monkeypatch fixture to set the state of
a test's fake Subscriptions API, the tests are not changed at all.
