+++ planet/cli/cli.py
@@ -19,7 +19,7 @@ import click
 
 import planet
 
-from . import auth, data, orders
+from . import auth, data, orders, subscriptions
 
 LOGGER = logging.getLogger(__name__)
 
@@ -78,3 +78,4 @@ def _configure_logging(verbosity):
 main.add_command(auth.auth)
 main.add_command(data.data)
 main.add_command(orders.orders)
+main.add_command(subscriptions.subscriptions)
